refactor: bring pagefind integration and update API reference page structure
