+++ package.json
@@ -9,16 +9,12 @@
     "preview": "astro preview",
     "astro": "astro",
     "upgrade": "pnpm dlx @astrojs/upgrade",
-    "index-api": "node scripts/algolia-api-indexer.js",
-    "test-urls": "node scripts/test-url-generation.js",
-    "test-indexing": "node scripts/test-indexing-dry-run.js",
-    "test-algolia": "node scripts/test-algolia-permissions.js"
+    "test-urls": "node scripts/test-url-generation.js"
   },
   "dependencies": {
     "@astrojs/netlify": "^6.3.4",
     "@astrojs/react": "^4.3.0",
     "@astrojs/starlight": "^0.34.3",
-    "@astrojs/starlight-docsearch": "^0.6.0",
     "@astrojs/starlight-tailwind": "^4.0.1",
     "@astrojs/tailwind": "^6.0.2",
     "@astrojs/vue": "^5.1.0",
@@ -34,7 +30,6 @@
     "@tailwindcss/vite": "^4.1.4",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.3",
-    "algoliasearch": "^5.25.0",
     "astro": "^5.8.0",
     "dotenv": "^16.5.0",
     "pnpm": "^10.10.0",
