+++ src/pages/apis.astro
@@ -1,29 +1,15 @@
 ---
-import "../styles/tailwind.css";
+import StarlightPage from '@astrojs/starlight/components/StarlightPage.astro';
 import ScalarReference from "../components/ScalarReference.vue";
+import ApiSearchIndex from "../components/ApiSearchIndex.astro";
 ---
 
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>API Reference</title>
-  </head>
-  <body>
-    <main>
-      <ScalarReference client:only="vue" />
-    </main>
-  </body>
-</html>
+<ScalarReference client:only="vue" />
 
-<style>
-  body {
-    margin: 0;
-    padding: 0;
-  }
-
-  main {
-    width: 100%;
-    height: 100vh;
-  }
-</style>
+<div style="display: none;">
+<StarlightPage frontmatter={{    description: 'Complete reference for all Scalekit API endpoints', tableOfContents: false, template: 'splash', title: 'API Reference' }}>
+  <div class="api-container">
+    <ApiSearchIndex />
+  </div>
+</StarlightPage>
+</div>
